feat: add sprites

+++ src/Main/GamePanel.java
@@ -1,36 +1,39 @@
 package Main;
 
-import java.awt.image.BufferedImage;
 import java.awt.*;
+import java.awt.image.BufferedImage;
 import java.awt.event.*;
+
 import javax.swing.JPanel;
 
 import GameState.GameStateManager;
 
 @SuppressWarnings("serial")
-public class GamePanel extends JPanel implements Runnable, KeyListener{
+public class GamePanel extends JPanel 
+	implements Runnable, KeyListener{
 	
-	// Dimension
+	// dimensions
 	public static final int WIDTH = 320;
 	public static final int HEIGHT = 240;
 	public static final int SCALE = 2;
 	
-	// Game Thread
+	// game thread
 	private Thread thread;
 	private boolean running;
 	private int FPS = 60;
-	private int targetTime = 1000 / FPS;
+	private long targetTime = 1000 / FPS;
 	
-	// Image
+	// image
 	private BufferedImage image;
 	private Graphics2D g;
 	
-	// Game State
+	// game state manager
 	private GameStateManager gsm;
 	
 	public GamePanel() {
 		super();
-		setPreferredSize(new Dimension(WIDTH * SCALE, HEIGHT * SCALE));
+		setPreferredSize(
+			new Dimension(WIDTH * SCALE, HEIGHT * SCALE));
 		setFocusable(true);
 		requestFocus();
 	}
@@ -45,15 +48,17 @@ public class GamePanel extends JPanel implements Runnable, KeyListener{
 	}
 	
 	private void init() {
+		
 		image = new BufferedImage(
-				WIDTH, HEIGHT,
-				BufferedImage.TYPE_INT_RGB
+					WIDTH, HEIGHT,
+					BufferedImage.TYPE_INT_RGB
 				);
 		g = (Graphics2D) image.getGraphics();
 		
 		running = true;
 		
 		gsm = new GameStateManager();
+		
 	}
 	
 	public void run() {
@@ -64,7 +69,7 @@ public class GamePanel extends JPanel implements Runnable, KeyListener{
 		long elapsed;
 		long wait;
 		
-		// Game Loop
+		// game loop
 		while(running) {
 			
 			start = System.nanoTime();
@@ -76,7 +81,7 @@ public class GamePanel extends JPanel implements Runnable, KeyListener{
 			elapsed = System.nanoTime() - start;
 			
 			wait = targetTime - elapsed / 1000000;
-			if (wait < 0) wait = 5;
+			if(wait < 0) wait = 5;
 			
 			try {
 				Thread.sleep(wait);
@@ -84,7 +89,9 @@ public class GamePanel extends JPanel implements Runnable, KeyListener{
 			catch(Exception e) {
 				e.printStackTrace();
 			}
+			
 		}
+		
 	}
 	
 	private void update() {
@@ -95,7 +102,9 @@ public class GamePanel extends JPanel implements Runnable, KeyListener{
 	}
 	private void drawToScreen() {
 		Graphics g2 = getGraphics();
-		g2.drawImage(image, 0, 0, WIDTH*SCALE, HEIGHT*SCALE, null);
+		g2.drawImage(image, 0, 0,
+				WIDTH * SCALE, HEIGHT * SCALE,
+				null);
 		g2.dispose();
 	}
 	
@@ -106,4 +115,21 @@ public class GamePanel extends JPanel implements Runnable, KeyListener{
 	public void keyReleased(KeyEvent key) {
 		gsm.keyReleased(key.getKeyCode());
 	}
+
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
